refactors organization of code for event listeners and event handlers
