+++ lib/food.js
@@ -1,6 +1,6 @@
 const $ = require('jquery');
-// const API = "https://quantified-self-rails-backend.herokuapp.com/api/v1"
-const API = 'https://quantified-self-express-api.herokuapp.com/api/v1'
+const API = "https://quantified-self-rails-backend.herokuapp.com/api/v1"
+    // const API = 'https://quantified-self-express-api.herokuapp.com/api/v1'
 
 class Food {
     constructor(food) {
@@ -62,6 +62,14 @@ class Food {
             data: { food: { calories: newCalories } }
         })
     }
+
+    // static update(id, newName, newCalories) {
+    //     return $.ajax({
+    //         url: API + `/foods/${id}`,
+    //         type: 'PUT',
+    //         data: { food: { name: newName, calories: newCalories } }
+    //     })
+    // }
 
     static showPopUp(el, id) {
         const popUp = `<div class='confirm-delete'>
