+++ lib/meal.js
@@ -1,7 +1,7 @@
 const $ = require('jquery');
 const Food = require('./food')
-    // const API = "https://quantified-self-rails-backend.herokuapp.com/api/v1"
-const API = 'https://quantified-self-express-api.herokuapp.com/api/v1'
+const API = "https://quantified-self-rails-backend.herokuapp.com/api/v1"
+    // const API = 'https://quantified-self-express-api.herokuapp.com/api/v1'
 
 
 class Meal {
